Move hook signature type to separate module
